requirements.txt
Django==1.5.5
Markdown==2.2.0
Pygments==1.5
South==0.7.6
argparse==1.2.1
#mercurial==2.4-rc
#psycopg2==2.4.5
uWSGI==1.4.5
wsgiref==0.1.2
#django-grappelli==2.4.3
django-debug-toolbar==0.9.4

# tracking is at https://hg.elevenbits.org/tracking
django-tracking==0.4.0

# treemenus is at https://hg.elevenbits.org/treemenus
django-treemenus==0.8.8-pre

fabric